app/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.2
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
deta==0.7
emoji-country-flag==1.2.3
fastapi==0.61.2
h11==0.11.0
httpcore==0.12.1
idna==2.10
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.11
requests==2.25.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.2
wrapt==1.12.1
yarl==1.6.2
